do tabbar
