+++ TeamNewsApp/BookmarksViewController.swift
@@ -5,4 +5,13 @@
 //  Created by Павел Широкий on 18.03.2024.
 //
 
-import Foundation
+import UIKit
+
+class BookmarksViewController: UIViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .cyan
+        
+    }
+}
