+++ TeamNewsApp/BrowseViewController.swift
@@ -5,4 +5,14 @@
 //  Created by Павел Широкий on 18.03.2024.
 //
 
-import Foundation
+import UIKit
+
+class BrowseViewController: UIViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .lightGray
+        
+    }
+    
+}
